fix : CLI functions moved to cli.py
