+++ setup.py
@@ -74,5 +74,5 @@ setup(
     license='MIT',
     entry_points={
         'console_scripts': [
-            'nava = nava.__main__:main',
+            'nava = nava.cli:main',
         ]})
